refactor(main.go): move getSubscriberIDs and addNewSubscriber functions to subscription.go file to improve code organization
